Enabled audio for Ubuntu
Copied the script and config from  overlay/etc/audio. The script will be
executed at boot-up or when pluginng in or removing the audio jack to write
registers to enable audio from jack. However, we need to remove audio.conf
to disable the auto switch since pusleaudio is not verified.

Change-Id: I3612ecb5586b8b8254a8a70d79c2e652f919eb61

Former-commit-id: 808654f347dcab6a51d06282d918033e97a93a90